refactor: change packet structure in sdk
