+++ include/tachimawari/control/sdk/packet/protocol_1/group_bulk_read.hpp
@@ -18,25 +18,29 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#ifndef TACHIMAWARI__CONTROL__SDK__UTILS__PROTOCOL_1__GROUP_BULK_READ_HPP_
-#define TACHIMAWARI__CONTROL__SDK__UTILS__PROTOCOL_1__GROUP_BULK_READ_HPP_
+#ifndef TACHIMAWARI__CONTROL__SDK__PACKET__PROTOCOL_1__GROUP_BULK_READ_HPP_
+#define TACHIMAWARI__CONTROL__SDK__PACKET__PROTOCOL_1__GROUP_BULK_READ_HPP_
 
 #include <map>
 #include <memory>
 #include <string>
 #include <vector>
 
-#include "tachimawari/control/sdk/utils/packet/group_bulk_read.hpp"
 #include "tachimawari/joint/model/joint.hpp"
+#include "tachimawari/joint/protocol_1/mx28_address.hpp"
 
 #include "dynamixel_sdk/dynamixel_sdk.h"
 
 namespace tachimawari::control::sdk::protocol_1
 {
 
-class GroupBulkRead : public tachimawari::control::sdk::GroupBulkRead
+class GroupBulkRead
 {
 public:
+  static void insert_all(
+    std::shared_ptr<std::map<uint8_t, GroupBulkRead>> bulk_data,
+    GroupBulkRead group_bulk_read);
+
   GroupBulkRead(
     dynamixel::PortHandler * port_handler,
     dynamixel::PacketHandler * packet_handler);
@@ -47,9 +51,22 @@ public:
     uint16_t data_length);
 
   void add(
-    const std::vector<tachimawari::joint::Joint> & joints) override;
+    const std::vector<tachimawari::joint::Joint> & joints);
+
+  int send();
+
+  int get(uint8_t id, uint16_t address, uint16_t data_length);
+
+  std::vector<uint8_t> get_parameters_id() const;
+
+  bool is_parameters_filled() const;
+
+private:
+  dynamixel::GroupBulkRead group_bulk_read;
+
+  std::vector<uint8_t> parameters_id;
 };
 
 }  // namespace tachimawari::control::sdk::protocol_1
 
-#endif  // TACHIMAWARI__CONTROL__SDK__UTILS__PROTOCOL_1__GROUP_BULK_READ_HPP_
+#endif  // TACHIMAWARI__CONTROL__SDK__PACKET__PROTOCOL_1__GROUP_BULK_READ_HPP_
