+++ src/tachimawari/control/sdk/packet/protocol_1/group_bulk_read.cpp
