+++ src/tachimawari/control/sdk/packet/protocol_1/group_sync_write.cpp
