+++ src/tachimawari/control/sdk/packet/protocol_2/group_bulk_read.cpp
