+++ src/tachimawari/control/sdk/packet/protocol_2/group_sync_write.cpp
